chore: 컴포넌트 폴더 내 모든 파일에 대해서 레지스트리 파일 생성하도록 수정

+++ packages/cli/src/scripts/build-registry.js
@@ -1,39 +1,47 @@
-import path from "path";
-import { promises as fs } from "fs";
+import { promises as fs, existsSync } from "fs";
 import { Project } from "ts-morph";
 import { filterExistingPath } from "../utils/filterExistingPath.js";
 
-const buildRegistry = async (component, componentPath, registryPath) => {
-  const absoluteComponentPath = path.resolve(componentPath);
+const buildRegistry = async (componentDir, registryPath) => {
+  if (!existsSync(componentDir) || !existsSync(registryPath)) {
+    console.error(`The path does not exist. Please try again.`);
+    process.exit(1);
+  }
 
   const project = new Project();
-  const sourceFile = project.addSourceFileAtPath(absoluteComponentPath);
+  const componentFiles = await fs.readdir(componentDir);
+  componentFiles.forEach(async (componentFile) => {
+    const componentFilePath = componentDir + "/" + componentFile;
+    const sourceFile = project.addSourceFileAtPath(componentFilePath);
 
-  const dependencies = [];
+    const dependencies = [];
 
-  sourceFile.getImportDeclarations().forEach((importDeclaration) => {
-    const module = importDeclaration.getModuleSpecifier().getLiteralValue();
+    sourceFile.getImportDeclarations().forEach((importDeclaration) => {
+      const module = importDeclaration.getModuleSpecifier().getLiteralValue();
 
-    if (filterExistingPath(module)) {
-      dependencies.push(module);
-    }
+      if (filterExistingPath(module)) {
+        dependencies.push(module);
+      }
+    });
+
+    const componentName = componentFile.split(".")[0];
+
+    const filePath = `${registryPath}/${componentName}.json`;
+
+    const fileContent = {
+      name: `${componentName}`,
+      dependencies,
+      files: [
+        {
+          name: `${componentName}.tsx`,
+          content: sourceFile.getText(),
+        },
+      ],
+    };
+    const stringifiedFileContent = JSON.stringify(fileContent);
+
+    await fs.writeFile(filePath, stringifiedFileContent);
   });
-
-  const filePath = `${registryPath}/${component}.json`;
-
-  const fileContent = {
-    name: `${component}`,
-    dependencies,
-    files: [
-      {
-        name: `${component}.tsx`,
-        content: sourceFile.getText(),
-      },
-    ],
-  };
-  const stringifiedFileContent = JSON.stringify(fileContent);
-
-  await fs.writeFile(filePath, stringifiedFileContent);
 };
 
 buildRegistry(process.argv[2], process.argv[3], process.argv[4]).catch(
